Dataset used to identify geography availability in the 5-year ACS Detailed Tables

Built-in dataset for use by load_variables() to identify the smallest geography at which 5-year ACS data are available

table: The ACS Table ID

geography: The smallest geography at which a given table is available for a given year

year: The endyear of the 5-year ACS dataset

data(acs5_geography)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 12228 rows and 3 columns.

Details

Dataset used to identify geography availability in the 5-year ACS Detailed Tables

Built-in dataset that includes information on the smallest geography at which 5-year ACS Detailed Tables data are available, by table, since 2011. This dataset is used internally by load_variables() to add a geography column when variables are retrieved for a 5-year ACS Detailed Tables dataset.