County geometry with Alaska and Hawaii shifted and re-scaled

Built-in dataset for use with shift_geo = TRUE

Dataset of US counties with Alaska and Hawaii shifted and re-scaled

data(county_laea)

data(county_laea)

Format

An object of class sf (inherits from data.frame) with 3143 rows and 2 columns.

Details

Dataset with county geometry for use when shifting Alaska and Hawaii

Built-in dataset for use with the shift_geo parameter, with the continental United States in a Lambert azimuthal equal area projection and Alaska and Hawaii counties and Census areas shifted and re-scaled. The data were originally obtained from the albersusa R package (https://github.com/hrbrmstr/albersusa).