Load variables from a decennial Census or American Community Survey dataset to search in R

Load variables from a decennial Census or American Community Survey dataset to search in R

load_variables(year, dataset, cache = FALSE)

Arguments

year
The year for which you are requesting variables. Either the year of the decennial Census, or the endyear for a 5-year ACS sample.

dataset
One of "sf1", "sf3", "acs1", "acs3", "acs5", "acs1/profile", "acs3/profile, "acs5/profile", "acs1/subject", "acs3/subject", or "acs5/subject".

cache
Whether you would like to cache the dataset for future access, or load the dataset from an existing cache. Defaults to FALSE.

Value

A tibble of variables from the requested dataset.

Examples

not_run({
  v15 <- load_variables(2015, "acs5", cache = TRUE)
  View(v15)
})