mig_recodes {tidycensus}	R Documentation

Dataset with Migration Flows characteristic recodes

Description

Built-in dataset for Migration Flows code label lookup.

characteristic: Characteristic variable name

code: Characteristic calue code

desc: Characteristic calue label

ordered: Whether or not recoded value should be ordered factor

Usage

data(mig_recodes)

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 120 rows and 4 columns.

Details

Dataset with Migration Flows characteristic recodes

Built-in dataset that is created from the Migration Flows API documentation. This dataset contains labels for the coded values returned by the Census API and is used when breakdown_labels = TRUE in get_flows.

[Package tidycensus version 1.1 Index]