moe_product {tidycensus}	R Documentation

Calculate the margin of error for a derived product

Description

Calculate the margin of error for a derived product

Usage

moe_product(est1, est2, moe1, moe2)

Arguments

est1
The first factor in the multiplication equation (an estimate)

est2
The second factor in the multiplication equation (an estimate)

moe1
The margin of error of the first factor

moe2
The margin of error of the second factor

Value

A margin of error for a derived product

[Package tidycensus version 1.1 Index]