moe_ratio {tidycensus}	R Documentation

Calculate the margin of error for a derived ratio

Description

Calculate the margin of error for a derived ratio

Usage

moe_ratio(num, denom, moe_num, moe_denom)

Arguments

num
The numerator involved in the ratio calculation (an estimate)

denom
The denominator involved in the ratio calculation (an estimate)

moe_num
The margin of error of the numerator

moe_denom
The margin of error of the denominator

Value

A margin of error for a derived ratio

[Package tidycensus version 1.1 Index]