pums_variables {tidycensus}	R Documentation

Dataset with PUMS variables and codes

Description

Built-in dataset for variable name and code label lookup. To access the data directly, issue the command data(pums_variables).

survey: acs1 or acs5

year: Year of data. For 5-year data, last year in range.

var_code: Variable name

var_label: Variable label

data_type: chr or num

level: housing or person

val_min: For numeric variables, the minimum value

val_max: For numeric variables, the maximum value

val_label: Value label

recode: Use labels to recode values

val_length: Length of value returned

val_na: Value of NA value returned by API (if known)

Usage

data(pums_variables)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 31759 rows and 12 columns.

Details

Dataset with PUMS variables and codes

Built-in dataset that is created from the Census PUMS data dictionaries. Use this dataset to lookup the names of variables to use in get_pums. This dataset also contains labels for the coded values returned by the Census API and is used when recode = TRUE in get_pums.

Because variable names and codes change from year to year, you should filter this dataset for the survey and year of interest. NOTE: 2017 - 2019 (acs1 and acs5) variables are available.

[Package tidycensus version 1.1 Index]