significance {tidycensus}	R Documentation

Evaluate whether the difference in two estimates is statistically significant.

Description

Evaluate whether the difference in two estimates is statistically significant.

Usage

significance(est1, est2, moe1, moe2, clevel = 0.9)

Arguments

est1
The first estimate.

est2
The second estimate

moe1
The margin of error of the first estimate

moe2
The margin of error of the second estimate

clevel
The confidence level. May by 0.9, 0.95, or 0.99

Value

TRUE if the difference is statistically signifiant, FALSE otherwise.

See Also

https://www.census.gov/content/dam/Census/library/publications/2018/acs/acs_general_handbook_2018_ch07.pdf

[Package tidycensus version 1.1 Index]