census_api_key {tidycensus}	R Documentation

Install a CENSUS API Key in Your .Renviron File for Repeated Use

Description

This function will add your CENSUS API key to your .Renviron file so it can be called securely without being stored in your code. After you have installed your key, it can be called any time by typing Sys.getenv("CENSUS_API_KEY") and can be used in package functions by simply typing CENSUS_API_KEY If you do not have an .Renviron file, the function will create on for you. If you already have an .Renviron file, the function will append the key to your existing file, while making a backup of your original file for disaster recovery purposes.

Usage

census_api_key(key, overwrite = FALSE, install = FALSE)

Arguments

key
The API key provided to you from the Census formated in quotes. A key can be acquired at http://api.census.gov/data/key_signup.html

overwrite
If this is set to TRUE, it will overwrite an existing CENSUS_API_KEY that you already have in your .Renviron file.

install
if TRUE, will install the key in your .Renviron file for use in future sessions. Defaults to FALSE.

Examples

## Not run: 
census_api_key("111111abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("CENSUS_API_KEY")

## End(Not run)

## Not run: 
# If you need to overwrite an existing key:
census_api_key("111111abc", overwrite = TRUE, install = TRUE)
# First time, relead your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("CENSUS_API_KEY")

## End(Not run)

[Package tidycensus version 1.1 Index]