Get data from the US Census Bureau Population Estimates APIs

Get data from the US Census Bureau Population Estimates APIs

get_estimates(
  geography,
  product = NULL,
  variables = NULL,
  breakdown = NULL,
  breakdown_labels = FALSE,
  year = 2019,
  state = NULL,
  county = NULL,
  time_series = FALSE,
  output = "tidy",
  geometry = FALSE,
  keep_geo_vars = FALSE,
  shift_geo = FALSE,
  key = NULL,
  show_call = FALSE,
  ...
)

Arguments

geography
The geography of your data.

product
The data product (optional). "population", "components" "housing", and "characteristics" are supported.

variables
A character string of requested variables to get specific variables from the population, components, and housing APIs.

breakdown
The population breakdown used when product = "characteristics". Acceptable values are "AGEGROUP", "RACE", "SEX", and "HISP", for Hispanic/Not Hispanic. These values can be combined in a vector, returning population estimates in the value column for all combinations of these breakdowns.

breakdown_labels
Whether or not to label breakdown elements returned when product = "characteristics". Defaults to FALSE.

year
The data year (defaults to 2018)

state
The state for which you are requesting data. State names, postal codes, and FIPS codes are accepted. Defaults to NULL.

county
The county for which you are requesting data. County names and FIPS codes are accepted. Must be combined with a value supplied to `state`. Defaults to NULL.

time_series
If TRUE, the function will return a time series of observations back to the decennial Census of 2010. The returned column is either "DATE", representing a particular estimate date, or "PERIOD", representing a time period (e.g. births between 2016 and 2017), and contains integers representing those values. Integer to date or period mapping is available at https://www.census.gov/data/developers/data-sets/popest-popproj/popest/popest-vars/2018.html.

output
One of "tidy" (the default) in which each row represents an enumeration unit-variable combination, or "wide" in which each row represents an enumeration unit and the variables are in the columns.

geometry
if FALSE (the default), return a regular tibble of ACS data. if TRUE, uses the tigris package to return an sf tibble with simple feature geometry in the `geometry` column.

keep_geo_vars
if TRUE, keeps all the variables from the Census shapefile obtained by tigris. Defaults to FALSE.

shift_geo
if TRUE, returns geometry with Alaska and Hawaii shifted for thematic mapping of the entire US.

key
Your Census API key. Obtain one at http://api.census.gov/data/key_signup.html. Can be stored in your .Renviron with census_api_key("YOUR KEY", install = TRUE)

show_call
if TRUE, display call made to Census API. This can be very useful in debugging and determining if error messages returned are due to tidycensus or the Census API. Copy to the API call into a browser and see what is returned by the API directly. Defaults to FALSE.

...
other keyword arguments

Value

A tibble, or sf tibble, of population estimates data